Move over the agents scheduling code
