+++ opencog/cogserver/server/CogServer.h
@@ -34,13 +34,9 @@
 #include <thread>
 
 #include <opencog/util/concurrent_queue.h>
-#include <opencog/cogserver/server/Agent.h>
-#include <opencog/cogserver/server/AgentRunnerBase.h>
-#include <opencog/cogserver/server/AgentRunnerThread.h>
 #include <opencog/cogserver/server/BaseServer.h>
 #include <opencog/cogserver/server/Module.h>
 #include <opencog/cogserver/server/NetworkServer.h>
-#include <opencog/cogserver/server/SystemActivityTable.h>
 #include <opencog/cogserver/server/Request.h>
 #include <opencog/cogserver/server/Registry.h>
 
@@ -51,26 +47,9 @@ namespace opencog
  */
 
 /**
- * This class implements a network and agent server. It provides shared
+ * This class implements a network server. It provides shared
  * network access to a given AtomSpace, together with some basic
- * capabilities to open a database, load C++ modules and to run
- * time-sliced processing "agents" called MindAgents.
- *
- * The module system is somewhat deprecated: python and scheme users
- * will typically find it more convenient to use the module system
- * provided by those langauges.  Thus, there is no requirement that new
- * C++ code must be wrapped into a CogServer module.
- *
- * Future versions of the CogServer might remove the module subsystem.
- *
- * The MindAgent subsystem is also somewhat obsolescent. It implements
- * a crude single-threaded cooperative time-slicing framework.  Most
- * users will simply find it easier to just create a new thread to run
- * whatever task that needs to be performed.
- *
- * Future versions of the CogServer might remove the agent subsystem.
- * Alternately, it might be useful to design a band-new thread
- * management subsystem. Maybe. Unclear.
+ * capabilities to load C++ extension modules.
  *
  * The most useful thing that the cogserver currently provides is
  * shared multi-user network access to a command line, from which
@@ -89,9 +68,6 @@ namespace opencog
  * python/scheme REPL shells, so there should be no issues with
  * multi-user access.
  *
- * Some (many?) of the C++ modules might not be thread-safe. 
- * None of the existing agents are thread-safe.
- *
  * -----------------------------------------------------------------
  * Implementation details:
  * 
@@ -106,21 +82,6 @@ namespace opencog
  * is called after the module has been instantiated and its meta-data 
  * has been filled.
  *
- * Cycles are handled by the server's main loop (method 'serverLoop').
- * Each cycle has a minimum duration controlled by the parameter
- * "SERVER_CYCLE_DURATION" (in seconds). At the start of every cycle,
- * the server processes the queued requests and then executes an
- * interaction of each scheduled agent. When all the agents are finished,
- * the server sleeps for the remaining time until the end of the cycle.
- *
- * Agent management is done through inheritance from the Registry<Agent>
- * class.  The agent registry API provides several methods to:
- *   1. register, unregister and list agents classes;
- *   2. create and destroy agent instances;
- *   3. start and stop agents.
- * We chose to wrap the register methods in the server class to avoid
- * conflicts with the other registry inheritance (Registry<Command>).
- *
  * Request management uses the same Registry base template as the agent
  * subsystem -- only, this time, using the Request base class. Thus, the
  * functionalities provided are very similar:
@@ -133,7 +94,7 @@ namespace opencog
  * That is, requests are assumed to run to completion in a small amount
  * of time, so that they don't block later requests.
  */
-class CogServer : public BaseServer, public Registry<Agent>, public Registry<Request>
+class CogServer : public BaseServer, public Registry<Request>
 {
     AtomSpace* _private_as;
 
@@ -149,28 +110,17 @@ protected:
         void*                   handle;
     } ModuleData;
     typedef std::map<const std::string, ModuleData> ModuleMap;
-    typedef std::unique_ptr<AgentRunnerThread> AgentRunnerThreadPtr;
 
     // Containers used to store references to the modules, requests
     // and agents
     ModuleMap modules;
     std::map<const std::string, Request*> requests;
 
-    long cycleCount;
-    bool running;
-    // Used to start and stop the Agents loop via shell commands
-    bool agentsRunning;
-
-    SimpleRunner agentScheduler;
-    std::vector<AgentRunnerThreadPtr> agentThreads;
-    std::map<std::string, AgentRunnerThread*> threadNameMap;
-
     std::mutex processRequestsMutex;
     concurrent_queue<Request*> requestQueue;
 
     NetworkServer* _networkServer;
-
-    SystemActivityTable _systemActivityTable;
+    bool _running;
 
 public:
 
@@ -196,24 +146,6 @@ public:
      *  public to be used in unit tests and for debug purposes only. */
     virtual void runLoopStep(void);
 
-    /** Customized server loop run. This method is called inside serverLoop
-     *  (between processing request queue and scheduled agents) and can be
-     *  overwritten by CogServer's subclasses in order to customize the
-     *  server loop behavior.
-     *
-     *  This method controls the execution of server cycles by returning
-     *  'true' if the server must run a cycle and 'false' if it must not.
-     *  By default it does nothing and returns true.
-     *
-     *  If EXTERNAL_TICK_MODE config parameter is enabled, serverLoop
-     *  will not go sleep at all. So, in this case, this method must be
-     *  in charge of going sleep when the server is idle to prevent
-     *  a busy-wait loop. */
-    virtual bool customLoopRun(void);
-
-    /** Returns the number of executed cycles so far */
-    virtual long getCycleCount(void);
-
     /** Terminates the main loop. The loop will be exited
      *  after the current interaction is finished. */
     virtual void stop(void);
@@ -225,9 +157,6 @@ public:
 
     /** Stops the network server and closes all the open server sockets. */
     virtual void disableNetworkServer(void);
-
-    /** Returns a reference to the system activity table instance */
-    virtual SystemActivityTable& systemActivityTable(void);
 
     /**** Module API ****/
     /** Loads a dynamic library/module. Takes the filename of the
@@ -259,48 +188,6 @@ public:
     virtual void loadModules(std::vector<std::string> module_paths =
                              std::vector<std::string>());
 
-    /**** Agent Registry API ****/
-    /** Register a new agent class/type. Takes the class' id and a
-     *  derived factory for this particular agent type. (note: the
-     *  caller owns the factory instance). */
-    virtual bool registerAgent(const std::string& id, AbstractFactory<Agent> const* factory);
-
-    /** Unregister an agent class/type. Takes the class' id. */
-    virtual bool unregisterAgent(const std::string& id);
-
-    /** Returns a list with the ids of all the registered agent classes. */
-    virtual std::list<const char*> agentIds(void) const;
-
-    /** Returns a list of all the currently running agent instances. */
-    virtual AgentSeq runningAgents(void);
-
-    /** Creates and returns a new instance of an agent of class 'id'.
-     *  If 'start' is true, then the agent will be automatically added
-     *  to the list of scheduled agents. */
-    virtual AgentPtr createAgent(const std::string& id,
-                                 const bool start = false);
-
-    /// Same as above, but returns the correct type.
-    template <typename T>
-    std::shared_ptr<T> createAgent(const bool start = false) {
-        return std::dynamic_pointer_cast<T>(createAgent(T::info().id, start));
-    }
-
-    /** Adds agent 'a' to the list of scheduled agents. */
-    virtual void startAgent(AgentPtr a, bool dedicated_thread = false,
-        const std::string &thread_name = "");
-
-    /** Removes agent 'a' from the list of scheduled agents. */
-    virtual void stopAgent(AgentPtr a);
-
-    /** Destroys all agents from class 'id' */
-    virtual void stopAllAgents(const std::string& id);
-
-    /** Starts running agents as part of the serverLoop (enabled by default) */
-    virtual void startAgentLoop(void);
-
-    /** Stops running agents as part of the serverLoop */
-    virtual void stopAgentLoop(void);
 
     /**** Request Registry API ****/
     /** Register a new request class/type. Takes the class id and a derived
